pyproject.toml
[project]
name = "sqlconnector"
version = "1.3.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Dan Ross", email = "daniel.ross@aop.com" }]
requires-python = ">=3.11"
dependencies = ["pyodbc>=5.2.0", "python-dotenv>=1.0.1", "sqlalchemy>=2.0.36"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
